Split API files
