+++ client/src/features/dataServices/dataServices.types.ts
@@ -59,64 +59,3 @@ export interface ResourcePoolsQueryParams {
   memoryRequest?: number;
   storageRequest?: number;
 }
-
-export interface CloudStorage {
-  storage: CloudStorageConfiguration;
-  sensitive_fields?: CloudStorageSensitiveFieldDefinition[];
-}
-
-export interface CloudStorageConfiguration {
-  configuration: Record<string, string | undefined>;
-  name: string;
-  private: boolean;
-  project_id: string;
-  source_path: string;
-  storage_id: string;
-  storage_type: string;
-  target_path: string;
-}
-
-export interface CloudStorageSensitiveFieldDefinition {
-  help: string;
-  name: string;
-}
-
-export interface GetCloudStorageForProjectParams {
-  project_id: string;
-}
-
-export type AddCloudStorageForProjectParams =
-  | AdvancedAddCloudStorageForProjectParams
-  | SimpleAddCloudStorageForProjectParams;
-
-export interface AdvancedAddCloudStorageForProjectParams {
-  configuration: Record<string, string | undefined>;
-  name: string;
-  private: boolean;
-  project_id: string;
-  source_path: string;
-  target_path: string;
-}
-
-export interface SimpleAddCloudStorageForProjectParams {
-  name: string;
-  private: boolean;
-  project_id: string;
-  storage_url: string;
-  target_path: string;
-}
-
-export interface UpdateCloudStorageParams {
-  configuration?: Record<string, string | null | undefined>;
-  name?: string;
-  private?: boolean;
-  project_id: string;
-  storage_id: string;
-  source_path?: string;
-  target_path?: string;
-}
-
-export interface DeleteCloudStorageParams {
-  project_id: string;
-  storage_id: string;
-}
