+++ client/src/features/session/components/options/SessionCloudStorageOptionV2.tsx
@@ -38,8 +38,8 @@ import { ErrorAlert } from "../../../../components/Alert";
 import { ExternalLink } from "../../../../components/ExternalLinks";
 import { Loader } from "../../../../components/Loader";
 import { Url } from "../../../../utils/helpers/url";
-import { useGetCloudStorageForProjectQuery } from "../../../dataServices/dataServicesApi";
 import { StateModelProject } from "../../../project/Project";
+import { useGetCloudStorageForProjectQuery } from "../../../project/projectCloudStorage.api";
 import { parseCloudStorageConfiguration } from "../../../project/utils/projectCloudStorage.utils";
 import { useGetNotebooksVersionsQuery } from "../../../versions/versionsApi";
 import { SessionCloudStorageV2 } from "../../startSessionOptions.types";
